Move static cache into parser engine instance
